+++ node/src/types/kernel/api/snippet/getSnippet/payload.d.ts
@@ -1,0 +1,35 @@
+/**
+ * Copyright (C) 2023 SiYuan Community
+ * 
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/**
+ * Get code snippet list
+ */
+export interface IPayload {
+    /**
+     * Get code snippets of a enabled state
+     */
+    enabled: number;
+    /**
+     * Get code snippets of a specified type
+     */
+    type: Type;
+}
+
+/**
+ * Get code snippets of a specified type
+ */
+export type Type = "js" | "css" | "all";